Added test cases for condition_specific, essential, and expression_utils
Created test cases for these files, specificall for the helper functions
in the condition_specific module, for the different aggregation
strategies in the essential module, and for the conversion methods in
expression utils. All tests are passing except those in the
condition_specific, since the test cases for the context specific
functions are currently in the process of being written.